agregamos validacion de busqueda en front

+++ backend/src/test/java/ar/edu/unq/turnero/service/TurnoServiceTest.kt
@@ -174,8 +174,8 @@ class TurnoServiceTest {
         turnoRecuperado = turnoService.recuperar(turnoId)
 
         Assert.assertEquals("Jorge Perez", turnoRecuperado!!.nombreYApellidoPaciente)
-        Assert.assertEquals(22345678, turnoRecuperado!!.dniPaciente)
-        Assert.assertEquals(1123456789, turnoRecuperado!!.telefonoPaciente)
+        Assert.assertEquals(22345678.toLong(), turnoRecuperado!!.dniPaciente)
+        Assert.assertEquals(1123456789.toLong(), turnoRecuperado!!.telefonoPaciente)
         Assert.assertEquals("jorgePerez@gmail.com", turnoRecuperado!!.emailPaciente)
     }
